Exit the Windows driver by pressing ESC anywhere on the desktop.
Prior to this changeset, to exit the Windows kbhell driver, one would
have to focus the window in which kbhell was launched and press ESC to
terminate the program. These changes allow the program to be terminated
by pressing ESC from anywhere on the desktop. After this change, the
Windows driver behavior is identical to that of the Linux driver.